Author: hforbert
redefined some arguments in linklists.F to be pointers to arrays instead
of arrays, as that's what we call these functions with and this can
produce runtime errors on some machines if the pointers are null.
Also fixed a small typo in format string of constraint output in
force_fields.F (this produced binary garbage in the output).


git-svn-id: svn://svn.code.sf.net/p/cp2k/code/trunk@853 bac0169d-1eea-4375-b5f6-f89d2ced9994